refactor: 위치 관련 패키지 분리
