+++ README.md
@@ -24,28 +24,29 @@ Position
 공통 내용
 - [ ] 각 말은 팀(Color)를 가지고 있다.
 - [ ] 도착지에 같은 팀이 있다면 이동할 수 없다.
-- [ ] 장기판 말 밖으로 이동할 수 없다.
+- [x] 장기판 말 밖으로 이동할 수 없다.
 
 비숍
-- [ ] 대각선 원하는만큼
+- [x] 대각선 원하는만큼
 - [ ] 뛰어넘기 불가
 
 왕
-- [ ] 상하좌우 + 대각선
+- [x] 상하좌우 + 대각선
 
 나이트
 - [ ] 상하좌우 한 칸 + 진행 방향으로 대각선 한 칸
 - [ ] 뛰어넘기 가능
 
 폰
-- [ ] 각 팀마다 진행방향으로 한칸 혹은 두 칸
+- [x] 각 팀마다 진행방향으로 한칸 이동 가능
+- [x] 처음 이동에는 두 칸 이동 가능
 - [ ] 상대편 말을 잡을 경우 대각선 이동 가능
 
 퀸
-- [ ] 상하좌우 원하는 만큼
-- [ ] 대각선 원하는 만큼
+- [x] 상하좌우 원하는 만큼
+- [x] 대각선 원하는 만큼
 - [ ] 뛰어넘기 불가
 
 룩
-- [ ] 상하좌우 원하는 만큼
+- [x] 상하좌우 원하는 만큼
 - [ ] 뛰어넘기 불가
