+++ src/main/java/chess/piece/Bishop.java
@@ -1,7 +1,7 @@
 package chess.piece;
 
-import chess.Color;
-import chess.Position;
+import chess.position.Color;
+import chess.position.Position;
 
 public class Bishop extends Piece {
