+++ src/main/java/chess/position/Color.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.position;
 
 public enum Color {
