+++ src/test/java/chess/ColumnTest.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.position.Column;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
