+++ src/test/java/chess/Fixtures.java
@@ -1,5 +1,9 @@
 package chess;
 
+import chess.position.Column;
+import chess.position.Position;
+import chess.position.Row;
+
 @SuppressWarnings("unused")
 public final class Fixtures {
