+++ src/test/java/chess/PositionTest.java
@@ -1,5 +1,7 @@
 package chess;
 
+import chess.position.Movement;
+import chess.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
