+++ src/test/java/chess/RowTest.java
@@ -1,5 +1,6 @@
 package chess;
 
+import chess.position.Row;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
